pyproject.toml
[project]
name = "pyrify"
version = "0.3.0"
description = "A CLI tool for database sanitization"
authors = [
    { name = "DataShades", email = "datashades@linkdigital.com.au"},
    { name = "Oleksandr", email = "mutantsan@gmail.com" }
]

dependencies = [
    "sqlalchemy>=2.0.40,<3.0.0",
    "psycopg2>=2.9.10,<3.0.0",
    "click>=8.1.8,<9.0.0",
    "PyYAML>=6.0.2,<7.0.0",
    "Faker>=37.1.0,<38.0.0",
    "PyMySQL>=1.1.1,<2.0.0",
]

[project.optional-dependencies]
dev = [
    "python-dotenv>=1.1.0,<2.0.0",
    "pytest>=8.3.5,<9.0.0",
    "ruff>=0.11.2,<1.0.0",
    "ipdb",
]

[project.scripts]
pyrify = "pyrify.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.isort]
known_self = "pyrify"
sections = "FUTURE,STDLIB,FIRSTPARTY,THIRDPARTY,SELF,LOCALFOLDER"
profile = "black"

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

# Same as Black.
line-length = 120
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"
